Refactor: move all constrained solvers into `torchmin/constrained` (TODO: replace minimize_constr.py file)
